+++ src/math.h
@@ -3,7 +3,7 @@
 
 double solveSimpleEquasion(double num1, char operator ,double num2);
 
-double mathPower(double num1, double num2);
+double mathPower(double num1, int num2);
 
 double mathModulo(double num1, double num2);
